Reverting system_max32672.h.

+++ MAX/Libraries/PeriphDrivers/Include/MAX32520/tmr.h
@@ -41,7 +41,6 @@
 #define _TMR_H_
 
 /* **** Includes **** */
-#include <stdint.h>
 #include "mxc_device.h"
 #include "tmr_regs.h"
 #include "mxc_sys.h"
@@ -218,7 +217,7 @@ void MXC_TMR_SetCount(mxc_tmr_regs_t *tmr, uint32_t cnt);
  * @param      tmr   The timer
  * @param[in]  us    microseconds to delay for
  */
-void MXC_TMR_Delay(mxc_tmr_regs_t *tmr, uint32_t us);
+void MXC_TMR_Delay(mxc_tmr_regs_t *tmr, unsigned long us);
 
 /**
  * @brief      Start a timer that will time out after a certain number of microseconds
@@ -226,7 +225,7 @@ void MXC_TMR_Delay(mxc_tmr_regs_t *tmr, uint32_t us);
  * @param      tmr   The timer
  * @param[in]  us    microseconds to time out after
  */
-void MXC_TMR_TO_Start(mxc_tmr_regs_t *tmr, uint32_t us);
+void MXC_TMR_TO_Start(mxc_tmr_regs_t *tmr, unsigned long us);
 
 /**
  * @brief      Check on time out timer
